app/requirements.txt
numpy>=1.26.4
pandas>=2.2.2
streamlit>=1.36.0
bibtexparser>=1.4.1
requests>=2.32.3
python-dotenv>=1.0.0
google>=3.0.0
pycurl>=7.45.3
certifi>=2024.2.2
arxiv>=2.1.3
elsapy>=0.5.1
crossrefapi>=1.5.1
alive-progress>=3.1.5
PyMuPDF>=1.24.9
google-generativeai>=0.7.2